+++ src/dev/components/Button.tsx
@@ -1,0 +1,9 @@
+import React, { FC } from 'react'
+
+export type ButtonProps = {
+  onClick: () => void
+}
+
+export const Button: FC<ButtonProps> = ({ onClick, children }) => {
+  return <button onClick={onClick}>{children}</button>
+}